Used two separate kernel classes instead `test_case` option
